reafactor:moved components to @components/misc and reimported them; changed navbar height and font size
