+++ app/user/[id]/wallet-history/page.tsx
@@ -1,37 +1,37 @@
-import dynamic from 'next/dynamic';
+// import dynamic from 'next/dynamic';
 
-import WorkInProgress from '@components/WorkInProgress';
+import WorkInProgress from '@components/misc/WorkInProgress';
 // import inflationWalletOverTimeValue from '@features/walletHistory/tools/inflationWalletOverTimeValue';
-import walletValueOverTime from '@features/walletHistory/tools/walletValueOverTime';
+// import walletValueOverTime from '@features/walletHistory/tools/walletValueOverTime';
 import { getUser } from '@src/api/UserApi';
 
-const WalletOverTimeCharts = dynamic(
-  () => import('@features/walletHistory/components/WalletOverTimeCharts'),
-);
+// const WalletOverTimeCharts = dynamic(
+//   () => import('@features/walletHistory/components/WalletOverTimeCharts'),
+// );
 
 const WalletHistoryPage = async () => {
   const user = await getUser();
   if (!user) return null;
 
-  const dailyWalletValues = await walletValueOverTime({
-    currencies: user.currencies,
-    quote_currency: user.quote_currency,
-    years: 1,
-  });
+  // const dailyWalletValues = await walletValueOverTime({
+  //   currencies: user.currencies,
+  //   quote_currency: user.quote_currency,
+  //   years: 1,
+  // });
 
   // const inflationDailyWalletValue = await inflationWalletOverTimeValue(
   //   dailyWalletValues,
   // );
-  const chartData = [
-    {
-      name: 'Wartość portfela',
-      data: dailyWalletValues.values,
-    },
-    // {
-    //   name: 'Realna wartość portfela',
-    //   data: inflationDailyWalletValue,
-    // },
-  ] as const;
+  // const chartData = [
+  //   {
+  //     name: 'Wartość portfela',
+  //     data: dailyWalletValues.values,
+  //   },
+  //   // {
+  //   //   name: 'Realna wartość portfela',
+  //   //   data: inflationDailyWalletValue,
+  //   // },
+  // ] as const;
 
   return (
     <WorkInProgress />
