--- src/components/segmentedSwitch/index.ts
@@ -1,2 +1,0 @@
-import SegmentedSwitch from "./SegmentedSwitch";
-export default SegmentedSwitch;
