+++ src/features/walletHistory/components/InflationOverMonthsTooltip.tsx
@@ -3,8 +3,8 @@ import { ReactElement } from 'react';
 import { DateTime } from 'luxon';
 import { TooltipProps } from 'recharts';
 
-import TooltipRowWrapper from '@components/TooltipRowWrapper';
-import TooltipWrapper from '@components/TooltipWrapper';
+import TooltipRowWrapper from '@components/tooltip/TooltipRowWrapper';
+import TooltipWrapper from '@components/tooltip/TooltipWrapper';
 import { CHART_THEME } from '@constants/chart';
 import { CustomTooltipProps } from '@interfaces/ICharts';
